risoluzione problemi

+++ Pista/model/Pista.py
@@ -9,12 +9,17 @@ class Pista():
         self.id = id
 
     def creaPista(self, disponibilita, id):
-         self.disponibilita = disponibilita
-         self.id = id
-         piste = []
-         with open('Pista/data/piste.pickle', "wb") as f:
-             pickle.dump(piste, f, pickle.HIGHEST_PROTOCOL)
-         return self
+        self.disponibilita = disponibilita
+        self.id = id
+        piste = []
+        if os.path.isfile('Pista/data/piste.pickle'):
+            with open('Pista/data/piste.pickle', "rb") as f:
+                piste = pickle.load(f)
+            piste.append(self)
+            print(piste)
+            with open('Pista/data/piste.pickle', "wb") as f:
+                pickle.dump(piste, f, pickle.HIGHEST_PROTOCOL)
+        return self
 
     def getPista(self):
         piste = []
@@ -32,5 +37,11 @@ class Pista():
         else:
             return None
 
-    def setDisponibilita(self, bool):
-        self.disponibilita = bool
+    def setDisponibilita(self, bool, id):
+        if os.path.isfile('Pista/data/piste.pickle'):
+            with open('Pista/data/piste.pickle', 'rb') as f:
+                piste = pickle.load(f)
+                pista = next((pista for pista in piste if str(pista.id) == str(id)), None)
+                pista.disponibilita = bool
+            with open('Pista/data/piste.pickle', 'wb') as f:
+                pickle.dump(piste, f, pickle.HIGHEST_PROTOCOL)
